This is a quick implementation for the `putchar(c)` just by replacing calls with `putc(c, stdio)`.
DISTRIBUTION STATEMENT “A”. (Approved for public release. Distribution is unlimited.)

This material is based upon work supported by DARPA under Contract No. HR001122C0047. Any opinions, findings and conclusions or recommendations expressed in this material are those of the author(s) and do not necessarily reflect the views of DARPA. DISTRIBUTION STATEMENT “A”. (Approved for public release. Distribution is unlimited.)